+++ src/LibHac/FsSrv/FsCreator/EmulatedGameCardFsCreator.cs
@@ -1,6 +1,7 @@
 ﻿using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
+using LibHac.Tools.Fs;
 
 namespace LibHac.FsSrv.FsCreator;
