+++ src/LibHac/Tools/Fs/Xci.cs
@@ -3,6 +3,7 @@ using LibHac.Common.Keys;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.FsSystem;
+using LibHac.Tools.Fs;
 
 namespace LibHac;
